Change CI scripts and fix Rust (without docs)
