+++ java/TypeDBExample.java
@@ -2,11 +2,9 @@
 // It is not intended for manual editing.
 package com.typedb.driver;
 
-import com.typedb.driver.TypeDB;
 import com.typedb.driver.api.DriverOptions;
 import com.typedb.driver.api.Credentials;
 import com.typedb.driver.api.Driver;
-import com.typedb.driver.api.QueryType;
 import com.typedb.driver.api.Transaction;
 import com.typedb.driver.api.answer.ConceptRow;
 import com.typedb.driver.api.answer.ConceptRowIterator;
@@ -21,13 +19,12 @@ import com.typedb.driver.common.exception.TypeDBDriverException;
 import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
-import java.util.Objects;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.stream.Collectors;
 
-public class TypeDBCoreExample {
+public class TypeDBExample {
     public void example() {
-        // Open a driver connection. Try-with-resources can be used for automatic driver connection management
+        // Open a driver connection. Specify your parameters if needed. Try-with-resources can be used for automatic driver connection management
         try (Driver driver = TypeDB.coreDriver(TypeDB.DEFAULT_ADDRESS, new Credentials("admin", "password"), new DriverOptions(false, null))) {
             // Create a database
             driver.databases().create("typedb");
