+++ rust/BUILD
@@ -112,20 +112,9 @@ rustdoc_to_adoc(
 )
 
 test_to_example(
-    name = "example_core",
-    input = "//rust/tests/integration/core:example.rs",
-    output = "rust/core_example.rs",
-    removed_lines = ["assert", "todo"],
-    changed_words = {
-        "example()": "typedb_example()",
-    },
-    header_comment_sign = "//",
-)
-
-test_to_example(
-    name = "example_cloud",
-    input = "//rust/tests/integration/cloud:example.rs",
-    output = "rust/cloud_example.rs",
+    name = "example",
+    input = "//rust/tests/integration:example.rs",
+    output = "rust/example.rs",
     removed_lines = ["assert", "todo"],
     changed_words = {
         "example()": "typedb_example()",
@@ -134,21 +123,11 @@ test_to_example(
 )
 
 update_markdown_example(
-    name = "readme_example_core",
-    input = ":core_example.rs",
+    name = "readme_example",
+    input = ":example.rs",
     output = ":README.md",
-    start_marker = "CORE_EXAMPLE_START_MARKER",
-    end_marker = "CORE_EXAMPLE_END_MARKER",
-    language_tag = "rust",
-    removed_header = "//",
-)
-
-update_markdown_example(
-    name = "readme_example_cloud",
-    input = ":cloud_example.rs",
-    output = ":README.md",
-    start_marker = "CLOUD_EXAMPLE_START_MARKER",
-    end_marker = "CLOUD_EXAMPLE_END_MARKER",
+    start_marker = "EXAMPLE_START_MARKER",
+    end_marker = "EXAMPLE_END_MARKER",
     language_tag = "rust",
     removed_header = "//",
 )
@@ -165,7 +144,7 @@ checkstyle_test(
         "**/Cargo.toml",
         "**/Cargo.lock",
         "docs/**/*.adoc",
-        "core_example.rs",
+        "example.rs",
         "cloud_example.rs",
     ]),
     license_type = "apache-header",
