+++ rust/tests/behaviour/steps/Cargo.toml
@@ -5,72 +5,72 @@
 features = {}
 
 [package]
-    name = "steps"
-    edition = "2021"
-    version = "0.0.0"
+	name = "steps"
+	edition = "2021"
+	version = "0.0.0"
 
 [lib]
-    path = "lib.rs"
+	path = "lib.rs"
 
 [dependencies]
 
-    [dependencies.tokio]
-        features = ["bytes", "default", "fs", "full", "io-std", "io-util", "libc", "macros", "mio", "net", "parking_lot", "process", "rt", "rt-multi-thread", "signal", "signal-hook-registry", "socket2", "sync", "time", "tokio-macros"]
-        version = "1.42.0"
-        default-features = false
+	[dependencies.tokio]
+		features = ["bytes", "default", "fs", "full", "io-std", "io-util", "libc", "macros", "mio", "net", "parking_lot", "process", "rt", "rt-multi-thread", "signal", "signal-hook-registry", "socket2", "sync", "time", "tokio-macros"]
+		version = "1.42.0"
+		default-features = false
 
-    [dependencies.smol]
-        features = []
-        version = "1.3.0"
-        default-features = false
+	[dependencies.smol]
+		features = []
+		version = "1.3.0"
+		default-features = false
 
-    [dependencies.regex]
-        features = ["default", "perf", "perf-backtrack", "perf-cache", "perf-dfa", "perf-inline", "perf-literal", "perf-onepass", "std", "unicode", "unicode-age", "unicode-bool", "unicode-case", "unicode-gencat", "unicode-perl", "unicode-script", "unicode-segment"]
-        version = "1.11.1"
-        default-features = false
+	[dependencies.regex]
+		features = ["default", "perf", "perf-backtrack", "perf-cache", "perf-dfa", "perf-inline", "perf-literal", "perf-onepass", "std", "unicode", "unicode-age", "unicode-bool", "unicode-case", "unicode-gencat", "unicode-perl", "unicode-script", "unicode-segment"]
+		version = "1.11.1"
+		default-features = false
 
-    [dependencies.cucumber]
-        features = ["default", "macros"]
-        version = "0.19.1"
-        default-features = false
+	[dependencies.cucumber]
+		features = ["default", "macros"]
+		version = "0.19.1"
+		default-features = false
 
-    [dependencies.async-std]
-        features = ["alloc", "async-attributes", "async-channel", "async-global-executor", "async-io", "async-lock", "attributes", "crossbeam-utils", "default", "futures-channel", "futures-core", "futures-io", "futures-lite", "gloo-timers", "kv-log-macro", "log", "memchr", "once_cell", "pin-project-lite", "pin-utils", "slab", "std", "wasm-bindgen-futures"]
-        version = "1.13.0"
-        default-features = false
+	[dependencies.async-std]
+		features = ["alloc", "async-attributes", "async-channel", "async-global-executor", "async-io", "async-lock", "attributes", "crossbeam-utils", "default", "futures-channel", "futures-core", "futures-io", "futures-lite", "gloo-timers", "kv-log-macro", "log", "memchr", "once_cell", "pin-project-lite", "pin-utils", "slab", "std", "wasm-bindgen-futures"]
+		version = "1.13.0"
+		default-features = false
 
-    [dependencies.macro_rules_attribute]
-        features = ["default"]
-        version = "0.2.0"
-        default-features = false
+	[dependencies.macro_rules_attribute]
+		features = ["default"]
+		version = "0.2.0"
+		default-features = false
 
-    [dependencies.typedb-driver]
-        path = "../../../../rust"
-        features = []
-        default-features = false
+	[dependencies.typedb-driver]
+		path = "../../../../rust"
+		features = []
+		default-features = false
 
-    [dependencies.futures]
-        features = ["alloc", "async-await", "default", "executor", "futures-executor", "std", "thread-pool"]
-        version = "0.3.31"
-        default-features = false
+	[dependencies.futures]
+		features = ["alloc", "async-await", "default", "executor", "futures-executor", "std", "thread-pool"]
+		version = "0.3.31"
+		default-features = false
 
-    [dependencies.chrono]
-        features = ["alloc", "android-tzdata", "clock", "default", "iana-time-zone", "js-sys", "now", "oldtime", "serde", "std", "wasm-bindgen", "wasmbind", "winapi", "windows-targets"]
-        version = "0.4.39"
-        default-features = false
+	[dependencies.chrono]
+		features = ["alloc", "android-tzdata", "clock", "default", "iana-time-zone", "js-sys", "now", "oldtime", "serde", "std", "wasm-bindgen", "wasmbind", "winapi", "windows-targets"]
+		version = "0.4.39"
+		default-features = false
 
-    [dependencies.uuid]
-        features = ["default", "fast-rng", "rng", "serde", "std", "v4"]
-        version = "1.11.0"
-        default-features = false
+	[dependencies.uuid]
+		features = ["default", "fast-rng", "rng", "serde", "std", "v4"]
+		version = "1.11.0"
+		default-features = false
 
-    [dependencies.itertools]
-        features = ["default", "use_alloc", "use_std"]
-        version = "0.10.5"
-        default-features = false
+	[dependencies.itertools]
+		features = ["default", "use_alloc", "use_std"]
+		version = "0.10.5"
+		default-features = false
 
-    [dependencies.serde_json]
-        features = ["alloc", "default", "indexmap", "preserve_order", "raw_value", "std"]
-        version = "1.0.135"
-        default-features = false
+	[dependencies.serde_json]
+		features = ["alloc", "default", "indexmap", "preserve_order", "raw_value", "std"]
+		version = "1.0.135"
+		default-features = false
