+++ rust/tests/integration/BUILD
@@ -17,7 +17,31 @@
 
 package(default_visibility = ["//visibility:public"])
 
+load("@rules_rust//rust:defs.bzl", "rust_test")
 load("@typedb_dependencies//tool/checkstyle:rules.bzl", "checkstyle_test")
+
+exports_files(
+    ["quickstart"],
+    visibility = ["//rust:__subpackages__"],
+)
+
+rust_test(
+    name = "test_example",
+    srcs = ["example.rs"],
+    deps = [
+        "//rust:typedb_driver",
+        "@crates//:async-std",
+        "@crates//:chrono",
+        "@crates//:futures",
+        "@crates//:itertools",
+        "@crates//:regex",
+        "@crates//:serde_json",
+        "@crates//:serial_test",
+        "@crates//:smol",
+        "@crates//:tokio",
+        "@crates//:uuid",
+    ],
+)
 
 checkstyle_test(
     name = "checkstyle",
