+++ rust/tests/integration/example.rs
@@ -1,6 +1,27 @@
-// This file is automatically generated.
-// It is not intended for manual editing.
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+// EXAMPLE START MARKER
 use futures::{StreamExt, TryStreamExt};
+// EXAMPLE END MARKER
+use serial_test::serial;
+// EXAMPLE START MARKER
 use typedb_driver::{
     answer::{
         concept_document::{Leaf, Node},
@@ -10,12 +31,32 @@ use typedb_driver::{
     Credentials, DriverOptions, Error, TransactionType, TypeDBDriver,
 };
 
-fn typedb_example() {
-    async_std::task::block_on(async {
-        // Open a driver connection
+// EXAMPLE END MARKER
 
-        let driver = TypeDBDriver::new_cloud(
-            &Vec::from([TypeDBDriver::DEFAULT_ADDRESS]),
+async fn cleanup() {
+    let driver = TypeDBDriver::new(
+        TypeDBDriver::DEFAULT_ADDRESS,
+        Credentials::new("admin", "password"),
+        DriverOptions::new(false, None).unwrap(),
+    )
+    .await
+    .unwrap();
+    if driver.databases().contains("typedb").await.unwrap() {
+        driver.databases().get("typedb").await.unwrap().delete().await.unwrap();
+    }
+}
+
+#[test]
+#[serial]
+// EXAMPLE START MARKER
+fn example() {
+    async_std::task::block_on(async {
+        // EXAMPLE END MARKER
+        cleanup().await;
+        // EXAMPLE START MARKER
+        // Open a driver connection. Specify your parameters if needed
+        let driver = TypeDBDriver::new(
+            TypeDBDriver::DEFAULT_ADDRESS,
             Credentials::new("admin", "password"),
             DriverOptions::new(false, None).unwrap(),
         )
@@ -25,6 +66,7 @@ fn typedb_example() {
         // Create a database
         driver.databases().create("typedb").await.unwrap();
         let database = driver.databases().get("typedb").await.unwrap();
+        assert_eq!(database.name(), "typedb");
 
         // Dropped transactions are closed
         {
@@ -61,6 +103,7 @@ fn typedb_example() {
         if answer.is_ok() && matches!(answer, QueryAnswer::Ok(_)) {
             println!("OK results do not give any extra interesting information, but they mean that the query is successfully executed!");
         }
+        assert!(answer.is_ok());
 
         // Commit automatically closes the transaction (don't forget to await the result!)
         transaction.commit().await.unwrap();
@@ -68,21 +111,26 @@ fn typedb_example() {
         // Open a read transaction to safely read anything without database modifications
         let transaction = driver.transaction(database.name(), TransactionType::Read).await.unwrap();
         let answer = transaction.query("match entity $x;").await.unwrap();
+        assert!(answer.is_row_stream());
 
         // Collect concept rows that represent the answer as a table
         let rows: Vec<ConceptRow> = answer.into_rows().try_collect().await.unwrap();
+        assert_eq!(rows.len(), 1);
         let row = rows.get(0).unwrap();
 
         // Retrieve column names to get concepts by index if the variable names are lost
         let column_names = row.get_column_names();
+        assert_eq!(column_names.len(), 1);
 
         let column_name = column_names.get(0).unwrap();
+        assert_eq!(column_name.as_str(), "x");
 
         // Get concept by the variable name (column name)
         let concept_by_name = row.get(column_name).unwrap().unwrap();
 
         // Get concept by the header's index
         let concept_by_index = row.get_index(0).unwrap().unwrap();
+        assert_eq!(concept_by_name, concept_by_index);
 
         // Check if it's an entity type
         if concept_by_name.is_entity_type() {
@@ -93,23 +141,37 @@ fn typedb_example() {
                 concept_by_index.get_label()
             );
         }
+        assert!(concept_by_name.is_entity_type());
+        assert!(concept_by_name.is_type());
+        assert_eq!(concept_by_name.get_label(), "person");
+        assert_ne!(concept_by_name.get_label(), "not person");
+        assert_ne!(concept_by_name.get_label(), "age");
 
         // Continue querying in the same transaction if needed
         let answer = transaction.query("match attribute $a;").await.unwrap();
+        assert!(answer.is_row_stream());
 
         // Concept rows can be used as a stream of results
         let mut rows_stream = answer.into_rows();
         while let Some(Ok(row)) = rows_stream.next().await {
             let mut column_names_iter = row.get_column_names().into_iter();
             let column_name = column_names_iter.next().unwrap();
+            assert_eq!(column_names_iter.next(), None);
 
             let concept_by_name = row.get(column_name).unwrap().unwrap();
+            assert!(concept_by_name.is_attribute_type());
+            assert!(concept_by_name.is_type());
+            assert!(concept_by_name.is_integer() || concept_by_name.is_string());
 
             // Check if it's an attribute type to safely retrieve its value type
             if concept_by_name.is_attribute_type() {
                 let label = concept_by_name.get_label();
                 let value_type = concept_by_name.try_get_value_type().unwrap();
                 println!("Defined attribute type's label: '{label}', value type: '{value_type}'");
+                assert!(value_type == ValueType::Integer || value_type == ValueType::String);
+                assert!(label == "age" || label == "name");
+                assert_ne!(concept_by_name.get_label(), "person");
+                assert_ne!(concept_by_name.get_label(), "person:age");
             }
 
             println!("It is also possible to just print the concept itself: '{}'", concept_by_name);
@@ -121,6 +183,7 @@ fn typedb_example() {
             .query("insert $z isa person, has age 10; $x isa person, has age 20, has name \"John\";")
             .await
             .unwrap();
+        assert!(answer.is_row_stream());
 
         // Insert queries also return concept rows
         let mut rows = Vec::new();
@@ -128,6 +191,7 @@ fn typedb_example() {
         while let Some(Ok(row)) = rows_stream.next().await {
             rows.push(row);
         }
+        assert_eq!(rows.len(), 1);
         let row = rows.get(0).unwrap();
 
         for column_name in row.get_column_names() {
@@ -140,6 +204,9 @@ fn typedb_example() {
 
         // It is possible to ask for the column names again
         let column_names = row.get_column_names();
+        assert_eq!(column_names.len(), 2);
+        assert!(column_names.contains(&"x".to_owned()));
+        assert!(column_names.contains(&"z".to_owned()));
 
         let x = row.get_index(column_names.iter().position(|r| r == "x").unwrap()).unwrap().unwrap();
         if let Some(iid) = x.try_get_iid() {
@@ -181,21 +248,31 @@ fn typedb_example() {
         // A match query can be used for concept row outputs
         let var = "x";
         let answer = transaction.query(format!("match ${} isa person;", var)).await.unwrap();
+        assert!(answer.is_row_stream());
 
         // Simple match queries always return concept rows
         let mut count = 0;
         let mut stream = answer.into_rows().map(|result| result.unwrap());
         while let Some(row) = stream.next().await {
             let x = row.get(var).unwrap().unwrap();
+            assert!(x.is_entity());
+            assert!(!x.is_entity_type());
+            assert!(!x.is_attribute());
+            assert!(!x.is_type());
+            assert!(x.is_instance());
+            assert_eq!(x.get_label(), "person");
             match x {
                 Concept::Entity(x_entity) => {
                     let x_type = x_entity.type_().unwrap();
+                    assert_eq!(x_type.label(), "person");
+                    assert_ne!(x_type.label(), "not person");
                     count += 1;
                     println!("Found a person {} of type {}", x, x_type);
                 }
                 _ => unreachable!("An entity is expected"),
             }
         }
+        assert_eq!(count, 4);
         println!("Total persons found: {}", count);
 
         // A fetch query can be used for concept document outputs with flexible structure
@@ -210,6 +287,7 @@ fn typedb_example() {
         };
         "#;
         let answer = transaction.query(fetch_query).await.unwrap();
+        assert!(answer.is_document_stream());
 
         // Fetch queries always return concept documents
         let mut count = 0;
@@ -244,7 +322,15 @@ fn typedb_example() {
             count += 1;
             println!("JSON representation of the fetched document:\n{}", document.into_json().to_string());
         }
+        assert_eq!(count, 5);
         println!("Total documents fetched: {}", count);
+        // EXAMPLE END MARKER
+    });
+
+    async_std::task::block_on(async {
+        cleanup().await;
+        // EXAMPLE START MARKER
         println!("More examples can be found in the API reference and the documentation.\nWelcome to TypeDB!");
     })
 }
+// EXAMPLE END MARKER
