+++ tool/test/start-community-server.sh
