+++ tool/test/stop-cluster-servers.sh
