+++ tool/test/stop-community-server.sh
